requirements.txt
runpod==1.7.13

diffusers==0.35.1
transformers==4.44.2
accelerate==0.33.0
huggingface_hub==0.25.2
safetensors==0.4.5
Pillow==10.4.0
numpy==1.26.4
